+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming
+### Changes
+* Json Trades 
+  * Added `tag_a` for a list of random items from the specified item tag
+  * For villager trades, added `low_priority` to replace modded trades 
+
 ## 4.16.1
 ### Changes
 * Apples now only drop from leaves in Spring